feat: added evaluation logging and end_run method

+++ src/tracker.py
@@ -26,7 +26,7 @@ class Model:
 
 class MLFlowTracker:
 
-    def __init__(self, client: MlflowClient, config: MLFlowConfig, number_of_models_to_track: int, min_or_max: Literal["min", "max"]) -> None:
+    def __init__(self, client: MlflowClient, config: MLFlowConfig, number_of_models_to_track: int, min_or_max: Literal["min", "max"] | None) -> None:
         self.artifact_dir = config.artifact_dir
         self.model_dir = config.models_dir
         self.client = client
@@ -171,7 +171,7 @@ class MLFlowTracker:
         except Exception as e:
             logger.warning(f"Artifact logging failed entirely: {e}")
             
-    def log_models(self, model, model_name: str, step: int = 0):
+    def log_model(self, model: nn.Module, mode: Literal["training", "evaluation"], model_name: str, step: int | None = None):
         self.model_name = f"{model_name}_{datetime.now().strftime("%Y-%m-%d_%H-%M-%S")}_{self.model_number}"
         model_info = mlflow.pytorch.log_model(pytorch_model=model, name=self.model_name, step=step) # type: ignore
         self.model_number += 1
@@ -181,13 +181,13 @@ class MLFlowTracker:
         pth_path = Path(self.artifact_dir) / self.model_dir / self.model_id / "artifacts/data/model.pth"
         pt_path = pth_path.with_name(self.model_name + ".pt")
         pth_path.rename(pt_path)
-
-        self.update_models_tracked()
-        if self.dropped_model_id is not None:
-            self.del_model_folder(self.dropped_model_id)
-            self.del_model_art(self.dropped_model_id)
-            self.client.delete_logged_model(self.dropped_model_id)
-            self.dropped_model_id = None
+        if mode == "training":
+            self.update_models_tracked()
+            if self.dropped_model_id is not None:
+                self.del_model_folder(self.dropped_model_id)
+                self.del_model_art(self.dropped_model_id)
+                self.client.delete_logged_model(self.dropped_model_id)
+                self.dropped_model_id = None
 
     def check_if_better(self, objective:float) -> bool:
 
@@ -231,8 +231,13 @@ class MLFlowTracker:
 
         if self.check_if_better(objective=model.best_val):
             self._current_best_val = model.best_val
-            self.log_models(model=model.model, model_name=model_name, step=step or 0)
+            self.log_model(model=model.model, mode="training", model_name=model_name, step=step or 0)
             self.log_metrics(metrics=model.metrics, step=step, model_id=self.model_id, artifact_path=self.model_name)
             self.log_metrics_artifact(metrics=model.metrics_art, save_metrics_for_model=True, artifact_path=self.model_name)
             self.log_config(config_path=model.config, save_config_for_model=True, save_as_parameters=True, artifact_path=self.model_name)
             
+    def log_evaluation(self, model: Model, model_name:str):
+        self.log_model(model=model.model, mode="evaluation", model_name=model_name)
+        self.log_metrics(metrics=model.metrics, model_id=self.model_id, artifact_path=self.model_name)
+        self.log_metrics_artifact(metrics=model.metrics_art, save_metrics_for_model=True, artifact_path=self.model_name)
+        self.log_config(config_path=model.config, save_config_for_model=True, save_as_parameters=True, artifact_path=self.model_name)
